chore: move pages directly inside pages folder
