+++ packages/frontend/src/__tests__/loginPage.js
@@ -1,7 +1,7 @@
 import { screen, render, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import UserContext from '../context/User'
-import LoginPage from '../pages/loginPage/loginPage'
+import LoginPage from '../pages/loginPage'
 
 jest.mock('react-router-dom', () => ({
     ...jest.requireActual('react-router-dom'),
