+++ packages/frontend/src/pages/newPage.tsx
@@ -2,11 +2,11 @@ import { useContext } from 'react'
 import { useHistory } from 'react-router-dom'
 import { observer } from 'mobx-react-lite'
 
-import UserContext from '../../context/User'
-import PostContext from '../../context/Post'
-import WritingField from '../../components/writingField'
-import BasicPage from '../basicPage/basicPage'
-import { DataType } from '../../constants'
+import UserContext from '../context/User'
+import PostContext from '../context/Post'
+import WritingField from '../components/writingField'
+import BasicPage from './basicPage'
+import { DataType } from '../constants'
 
 const NewPage = () => {
     const history = useHistory()
