+++ packages/frontend/src/router.tsx
@@ -5,15 +5,15 @@ import useLocalStorage from './hooks/useLocalStorage'
 import * as Constants from './constants'
 
 import Header from './layout/header/header'
-import MainPage from './pages/mainPage/mainPage'
-import PostPage from './pages/postPage/postPage'
-import UserPage from './pages/userPage/userPage'
-import LoginPage from './pages/loginPage/loginPage'
-import SignupPage from './pages/signupPage/signupPage'
-import NewPage from './pages/newPage/newPage'
-import FeedbackPage from './pages/feedbackPage/feedbackPage'
-import AdminPage from './pages/adminPage/adminPage'
-import SettingPage from './pages/settingPage/settingPage'
+import MainPage from './pages/mainPage'
+import PostPage from './pages/postPage'
+import UserPage from './pages/userPage'
+import LoginPage from './pages/loginPage'
+import SignupPage from './pages/signupPage'
+import NewPage from './pages/newPage'
+import FeedbackPage from './pages/feedbackPage'
+import AdminPage from './pages/adminPage'
+import SettingPage from './pages/settingPage'
 
 import { WebContext } from './context/WebContext'
 import Favicon from 'react-favicon'
